+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## next
+
+- Added shortcuts in search input:
+  - `Enter` – select next matched
+  - `Shift+Enter` – select previous matched
+  - `Escape` – clear search input
+
 ## 0.5.0 (October 10, 2021)
 
 - Added short info block for selected fiber